examples/web_scraper_example.py
"""Example provider implementation for web scraping."""
import requests
from bs4 import BeautifulSoup
from typing import List, Dict
import logging
from datetime import datetime
import sys
import os

# Add parent directory to path for imports
sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))

from providers.base_provider import BaseProvider

logger = logging.getLogger(__name__)


class WebScraperProvider(BaseProvider):
    """
    Example provider that scrapes web pages.
    
    This is a template - adapt to specific site structure.
    """
    
    def __init__(self, base_url: str, name: str = "web_scraper", enabled: bool = True):
        """
        Initialize web scraper provider.
        
        Args:
            base_url: Base URL to scrape
            name: Provider identifier
            enabled: Whether provider is active
        """
        super().__init__(name, enabled)
        self.base_url = base_url
        self.session = requests.Session()
        self.session.headers.update({
            "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
        })
    
    def search(self, query: str, count: int = 10, **kwargs) -> List[Dict]:
        """
        Scrape search results from web page.
        
        Args:
            query: Search terms
            count: Maximum results
            **kwargs: Additional parameters
            
        Returns:
            List of standardized results
        """
        # Check cache first
        cached = self.get_cached_results(query)
        if cached:
            logger.info("Returning cached results for: %s", query)
            return cached[:count]
        
        try:
            # Build search URL (adapt to target site)
            search_url = f"{self.base_url}/search?q={query}"
            
            response = self.session.get(search_url, timeout=10)
            response.raise_for_status()
            
            # Parse HTML
            soup = BeautifulSoup(response.text, 'html.parser')
            results = self._parse_results(soup)
            
            # Normalize and validate
            normalized = []
            for item in results:
                if self.validate(item):
                    normalized.append(self.normalize(item))
            
            # Cache results
            self.cache_results(query, normalized)
            
            logger.info("Found %d results for: %s", len(normalized), query)
            return normalized[:count]
            
        except requests.RequestException as e:
            logger.error("Request failed for %s: %s", query, e)
            return []
        except Exception as e:
            logger.error("Parsing failed: %s", e)
            return []
    
    def _parse_results(self, soup: BeautifulSoup) -> List[Dict]:
        """
        Parse results from HTML soup.
        
        Adapt this method to target site's structure.
        """
        results = []
        
        # Example: Find result containers (adapt selectors)
        result_elements = soup.select('.result-item')
        
        for element in result_elements:
            try:
                # Extract data (adapt to site structure)
                title_elem = element.select_one('.title')
                link_elem = element.select_one('a')
                desc_elem = element.select_one('.description')
                
                if not (title_elem and link_elem):
                    continue
                
                result = {
                    "id": link_elem.get('href', ''),
                    "title": title_elem.get_text(strip=True),
                    "link": link_elem.get('href', ''),
                    "description": desc_elem.get_text(strip=True) if desc_elem else "",
                    "source": self.name,
                    "timestamp": datetime.now().isoformat(),
                    "metadata": {
                        # Add site-specific metadata
                        "scraped_from": self.base_url
                    }
                }
                
                results.append(result)
                
            except Exception as e:
                logger.debug("Failed to parse element: %s", e)
                continue
        
        return results
    
    def normalize(self, raw_item: Dict) -> Dict:
        """Normalize scraped data to standard format."""
        normalized = super().normalize(raw_item)
        
        # Ensure absolute URLs
        if normalized["link"] and not normalized["link"].startswith("http"):
            normalized["link"] = f"{self.base_url}{normalized['link']}"
        
        return normalized


# Example usage
if __name__ == "__main__":
    # Test the provider
    provider = WebScraperProvider(base_url="https://example.com")
    results = provider.search("python developer", count=5)
    
    print(f"Found {len(results)} results:")
    for result in results:
        print(f"- {result['title']}")
        print(f"  {result['link']}")
        print()
